+++ src/components/radio/radio.styles.ts
@@ -98,6 +98,7 @@ export default css`
   }
 
   .radio__label {
+    color: var(--sl-input-label-color);
     line-height: var(--sl-toggle-size);
     margin-inline-start: 0.5em;
     user-select: none;