+++ docs/resources/changelog.md
@@ -6,7 +6,7 @@ Components with the <sl-badge variant="warning" pill>Experimental</sl-badge> bad
 
 _During the beta period, these restrictions may be relaxed in the event of a mission-critical bug._ 🐛
 
-## Next
+## 2.0.0-beta.77
 
 - Added styles to required form controls so they show an asterisk next to the label by default
 - Added the `--sl-input-required-content` design token